What I did in 30 minutes
I am using my own source code, with bar and table classes. To make the client going to the bar, I am still asking via terminal data from the user. In this case, I used the 30 minutes to get correctly the phone number to the reserve, because I am thinking in the case of doubles names. Then, I will give to the user the option of make a reserve or going to the bar, in which case the selected bar will ask the name. If that name is on the table array, it will sit the user in that table.